fixed downcast warning
git-svn-id: https://svn-i2.informatik.rwth-aachen.de/repos/libalf/trunk@1383 ee1db7db-6c55-0410-a809-e4d5348c71a5

+++ include/serialize.h
@@ -131,7 +131,7 @@ inline                                  std::basic_string<int32_t> serialize(dou
 
 	significand = std::frexp(a, &exponent);
 
-	sig2 = significand * (2 << 30);
+	sig2 = (int)significand * (2 << 30);
 
 	ret += serialize(exponent);
 	ret += serialize(sig2);
